updating multiexecutor fix fixed a bug in killAll

+++ trunk/src/com/flaptor/clusterfest/controlling/ControllerNodeDescriptor.java
@@ -77,9 +77,11 @@ public class ControllerNodeDescriptor extends ModuleNodeDescriptor {
 	 */
 	public void kill() throws IOException {
 		updateState();
-		Triad<Integer, String, String> rv = CommandUtil.remoteSSHCommand(getNodeDescriptor().getHost(), -1, "cd "+getNodeDescriptor().getInstallDir() + "\n./stop.sh", 30000);
-		if (rv.first() == 0) return;
-		else throw new IOException("could not kill remote host. error code: " + rv.first() + " - " + rv.second() + " - " + rv.third()); 
+        if (state != ControllerNodeState.STOPPED) {
+    		Triad<Integer, String, String> rv = CommandUtil.remoteSSHCommand(getNodeDescriptor().getHost(), -1, "cd "+getNodeDescriptor().getInstallDir() + "\n./stop.sh", 30000);
+    		if (rv.first() == 0) return;
+    		else throw new IOException("could not kill remote host. error code: " + rv.first() + " - " + rv.second() + " - " + rv.third());
+        }
 	}
 
 	/**
